Create robber
For this question, I was given an array of numbers, and I had to return the maximum sum on numbers where none were adjacent to each other. I was also told to act as if the list was circular, so the first and last entries were adjacent.

I found a great explanation in the discussions for this question that I based my solution on. To save time and space, I could just compare the previous entry to the one before that entry plus the value of whatever index I'm currently on, and pick the bigger value to be my new sum. This allowed me to keep track of the total without creating a new list.